requirements.txt
discord==1.7.3
discord.py==1.7.3
requests==2.27.1
typing_extensions==4.1.0
names==0.3.0
py-cord==2.0.0b1
matplotlib==3.3.4
librosa==0.9.1
requests-toolbelt==0.7.0
scapy==2.4.4
pydub==0.25.1
tinyscript==1.24.14
stegoveritas==1.9
stego-lsb==1.3.1
stegpy==0.0.1
oletools==0.60
cryptography==36.0.2
pyUnicodeSteganography==0.0.1
emoji==1.6.3
stringcheese==2.1
name-that-hash==1.10.0
pwntools==4.7.0
r2pipe==1.6.5
factordb-pycli==1.3.0
tqdm==4.57.0
z3-solver==4.8.12.0
pypcapkit==0.15.5
scipy==1.5.4
cipheycore==0.3.2
ciphey==5.14.0
cipheydists==0.3.35
gmpy2
sage